// src/lib/data/artists.generated.ts
export const ARCHIVE_ARTISTS = [
  {
    "slug": "333",
    "name": "333, Joeski"
  },
  {
    "slug": "333-3ski",
    "name": "333 (3SKI), Joeski"
  },
  {
    "slug": "afrikali",
    "name": "Afrikali"
  },
  {
    "slug": "aldo-cadiz",
    "name": "Aldo Cadiz"
  },
  {
    "slug": "alex-davis",
    "name": "Alex Davis, Herald"
  },
  {
    "slug": "alex-sosa",
    "name": "Alex Sosa"
  },
  {
    "slug": "alexander-louis",
    "name": "Alexander:Louis"
  },
  {
    "slug": "alfred-azzetto",
    "name": "Alfred Azzetto"
  },
  {
    "slug": "benji-candelario",
    "name": "Benji Candelario"
  },
  {
    "slug": "brooklyn-pirates",
    "name": "Brooklyn Pirates, Liberty"
  },
  {
    "slug": "chris-fortier",
    "name": "Chris Fortier"
  },
  {
    "slug": "cricco-castelli",
    "name": "Cricco Castelli"
  },
  {
    "slug": "cultural-vibe",
    "name": "Cultural Vibe Feat. Keith Thompson"
  },
  {
    "slug": "david-herrero",
    "name": "David Herrero"
  },
  {
    "slug": "de-no",
    "name": "de No, Joeski"
  },
  {
    "slug": "denney",
    "name": "Denney"
  },
  {
    "slug": "dj-chus",
    "name": "DJ Chus, Joeski"
  },
  {
    "slug": "dj-taco",
    "name": "DJ Taco"
  },
  {
    "slug": "el-jibaro",
    "name": "El Jibaro"
  },
  {
    "slug": "funkyloco",
    "name": "Funkyloco"
  },
  {
    "slug": "george-privatti",
    "name": "George Privatti"
  },
  {
    "slug": "green-velvet",
    "name": "Green Velvet, Joeski"
  },
  {
    "slug": "harry-romero",
    "name": "Harry Romero"
  },
  {
    "slug": "hipp-e",
    "name": "Hipp-E"
  },
  {
    "slug": "hr-ski",
    "name": "HR&Ski"
  },
  {
    "slug": "ismael-rivas",
    "name": "Ismael Rivas"
  },
  {
    "slug": "j",
    "name": "J"
  },
  {
    "slug": "javi-lopez",
    "name": "Javi Lopez"
  },
  {
    "slug": "jesante",
    "name": "Jesante, Joeski"
  },
  {
    "slug": "joeski",
    "name": "Joeski"
  },
  {
    "slug": "joeski-javi-lopez-ruben-zurita",
    "name": "Joeski Javi Lopez Ruben Zurita"
  },
  {
    "slug": "jose-m",
    "name": "Jose M., Matheo Velez, TacoMan"
  },
  {
    "slug": "juanito",
    "name": "Juanito"
  },
  {
    "slug": "k-ax",
    "name": "K'áax"
  },
  {
    "slug": "kax",
    "name": "Káax"
  },
  {
    "slug": "ki-creighton",
    "name": "Ki Creighton, Makanan"
  },
  {
    "slug": "le-vinyl",
    "name": "Le Vinyl"
  },
  {
    "slug": "leoesco",
    "name": "Leoesco"
  },
  {
    "slug": "level-groove",
    "name": "Level Groove"
  },
  {
    "slug": "lez",
    "name": "&lez"
  },
  {
    "slug": "los-bandidos",
    "name": "Los Bandidos"
  },
  {
    "slug": "lucio-grandi",
    "name": "Lucio Grandi"
  },
  {
    "slug": "m-j",
    "name": "M&J"
  },
  {
    "slug": "macondo",
    "name": "Macondo"
  },
  {
    "slug": "martin-aquino",
    "name": "Martin Aquino"
  },
  {
    "slug": "max-martinez",
    "name": "Max Martinez"
  },
  {
    "slug": "mert-yucel",
    "name": "Mert Yucel"
  },
  {
    "slug": "miguel-lobo",
    "name": "Miguel Lobo"
  },
  {
    "slug": "mikee-athens",
    "name": "Mikee (Athens)"
  },
  {
    "slug": "moses-mcclean",
    "name": "Moses McClean"
  },
  {
    "slug": "mr-ski",
    "name": "Mr. Ski"
  },
  {
    "slug": "nick",
    "name": "Nick"
  },
  {
    "slug": "penn",
    "name": "Penn"
  },
  {
    "slug": "robb-swinga",
    "name": "Robb Swinga"
  },
  {
    "slug": "robert-garcia",
    "name": "Robert Garcia"
  },
  {
    "slug": "serious-intention",
    "name": "Serious Intention"
  },
  {
    "slug": "shaboom-allstars",
    "name": "Shaboom Allstars"
  },
  {
    "slug": "stereo-mutants",
    "name": "Stereo Mutants"
  },
  {
    "slug": "sydney-blu",
    "name": "Sydney Blu"
  },
  {
    "slug": "tony-senghore",
    "name": "Tony Senghore"
  },
  {
    "slug": "tripio-x",
    "name": "Tripio X"
  },
  {
    "slug": "vangelis-kostoxenakis",
    "name": "Vangelis Kostoxenakis"
  }
] as const;